refactor(materializer): extract conversation-stability helpers (file #3)
Lift the pure conversation-patch stability predicates + collectors (isStable*,
collect*, areSceneEntryListsEquivalent, materializeVisibleInteractionEntries)
into conversation-stability.ts. Net 74/74, check 0.
