+++ packages/desktop/src/lib/acp/session-state/agent-panel-graph-materializer.ts
@@ -82,6 +82,16 @@ import {
 	createTruncatedSceneEntryRowIndex,
 } from "./scene-entry-row-index.js";
 import {
+	areSceneEntryListsEquivalent,
+	collectAppendedInteractions,
+	collectAppendedTranscriptEntries,
+	collectStableTranscriptPatchedEntriesByIndex,
+	isStableTranscriptAppend,
+	isStableTranscriptPatchAndAppend,
+	isStableTranscriptTruncation,
+	materializeVisibleInteractionEntries,
+} from "./conversation-stability.js";
+import {
 	interactionSceneEntryId,
 	materializeTranscriptEntry,
 	questionInteractionToSceneEntry,
@@ -1757,141 +1767,6 @@ function materializeMarkedInteractionPatchedConversation(
 		},
 		sceneEntryRowIndex,
 	};
-}
-
-function materializeVisibleInteractionEntries(
-	graph: AgentPanelCanonicalSource,
-	sceneEntryRowIndex: ReadonlyMap<string, number>
-): AgentPanelSceneEntryModel[] {
-	const entries: AgentPanelSceneEntryModel[] = [];
-	for (const interaction of graph.interactions) {
-		const entry = questionInteractionToSceneEntry(interaction, graph);
-		if (entry === null || sceneEntryRowIndex.has(entry.id)) {
-			continue;
-		}
-		entries.push(entry);
-	}
-	return entries;
-}
-
-function areSceneEntryListsEquivalent(
-	left: readonly AgentPanelSceneEntryModel[],
-	right: readonly AgentPanelSceneEntryModel[]
-): boolean {
-	if (left.length !== right.length) {
-		return false;
-	}
-	return left.every((entry, index) => areSceneEntriesEquivalent(entry, right[index]));
-}
-
-function isStableTranscriptAppend(
-	previousEntries: readonly TranscriptEntry[],
-	nextEntries: readonly TranscriptEntry[]
-): boolean {
-	if (nextEntries.length < previousEntries.length) {
-		return false;
-	}
-
-	for (let index = 0; index < previousEntries.length; index += 1) {
-		if (nextEntries[index] !== previousEntries[index]) {
-			return false;
-		}
-	}
-
-	return true;
-}
-
-function isStableTranscriptPatchAndAppend(
-	previousEntries: readonly TranscriptEntry[],
-	nextEntries: readonly TranscriptEntry[]
-): boolean {
-	if (nextEntries.length <= previousEntries.length) {
-		return false;
-	}
-
-	let sawPatchedPrefixEntry = false;
-	for (let index = 0; index < previousEntries.length; index += 1) {
-		const previousEntry = previousEntries[index];
-		const nextEntry = nextEntries[index];
-		if (previousEntry === undefined || nextEntry === undefined) {
-			return false;
-		}
-		if (previousEntry.entryId !== nextEntry.entryId) {
-			return false;
-		}
-		if (previousEntry !== nextEntry) {
-			sawPatchedPrefixEntry = true;
-		}
-	}
-	return sawPatchedPrefixEntry;
-}
-
-function isStableTranscriptTruncation(
-	previousEntries: readonly TranscriptEntry[],
-	nextEntries: readonly TranscriptEntry[]
-): boolean {
-	if (nextEntries.length >= previousEntries.length) {
-		return false;
-	}
-
-	for (let index = 0; index < nextEntries.length; index += 1) {
-		if (nextEntries[index] !== previousEntries[index]) {
-			return false;
-		}
-	}
-
-	return true;
-}
-
-function collectStableTranscriptPatchedEntriesByIndex(
-	previousEntries: readonly TranscriptEntry[],
-	nextEntries: readonly TranscriptEntry[]
-): ReadonlyMap<number, TranscriptEntry> | null {
-	let patches: Map<number, TranscriptEntry> | null = null;
-	for (let index = 0; index < previousEntries.length; index += 1) {
-		const previousEntry = previousEntries[index];
-		const nextEntry = nextEntries[index];
-		if (previousEntry === undefined || nextEntry === undefined) {
-			return null;
-		}
-		if (previousEntry.entryId !== nextEntry.entryId) {
-			return null;
-		}
-		if (previousEntry === nextEntry) {
-			continue;
-		}
-		patches ??= new Map<number, TranscriptEntry>();
-		patches.set(index, nextEntry);
-	}
-	return patches;
-}
-
-function collectAppendedTranscriptEntries(
-	entries: readonly TranscriptEntry[],
-	startIndex: number
-): readonly TranscriptEntry[] {
-	const appendedEntries: TranscriptEntry[] = [];
-	for (let index = startIndex; index < entries.length; index += 1) {
-		const entry = entries[index];
-		if (entry !== undefined) {
-			appendedEntries.push(entry);
-		}
-	}
-	return appendedEntries;
-}
-
-function collectAppendedInteractions(
-	interactions: readonly InteractionSnapshot[],
-	startIndex: number
-): readonly InteractionSnapshot[] {
-	const appendedInteractions: InteractionSnapshot[] = [];
-	for (let index = startIndex; index < interactions.length; index += 1) {
-		const interaction = interactions[index];
-		if (interaction !== undefined) {
-			appendedInteractions.push(interaction);
-		}
-	}
-	return appendedInteractions;
 }
 
 function materializeOperationPatchedConversation(
